+++ src/main/java/br/com/pedrolg/projeto/GrupocarecaspringApplication.java
@@ -1,0 +1,13 @@
+package br.com.pedrolg.projeto;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GrupocarecaspringApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(GrupocarecaspringApplication.class, args);
+	}
+
+}
