+++ src/main/java/br/com/pedrolg/projeto/controller/ClienteController.java
@@ -1,0 +1,50 @@
+package br.com.pedrolg.projeto.controller;
+
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import br.com.pedrolg.projeto.entity.Cliente;
+import br.com.pedrolg.projeto.service.ClienteService;
+
+@RestController
+@RequestMapping("/clientes")
+public class ClienteController {
+
+    @Autowired
+    private ClienteService clienteService;
+
+    @GetMapping
+    public ResponseEntity<List<Cliente>> listarClientes() {
+        List<Cliente> clientes = clienteService.listarClientes();
+        return new ResponseEntity<>(clientes, HttpStatus.OK);
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<Cliente> encontrarClientePorId(@PathVariable int id) {
+        return clienteService.encontrarClientePorId(id)
+                .map(cliente -> new ResponseEntity<>(cliente, HttpStatus.OK))
+                .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
+    }
+
+    @PostMapping
+    public ResponseEntity<Cliente> salvarCliente(@RequestBody Cliente cliente) {
+        Cliente novoCliente = clienteService.salvarCliente(cliente);
+        return new ResponseEntity<>(novoCliente, HttpStatus.CREATED);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deletarCliente(@PathVariable int id) {
+        clienteService.deletarCliente(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
+}
