+++ src/main/java/br/com/pedrolg/projeto/controller/ItemController.java
@@ -1,0 +1,50 @@
+package br.com.pedrolg.projeto.controller;
+
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import br.com.pedrolg.projeto.entity.Item;
+import br.com.pedrolg.projeto.service.ItemService;
+
+@RestController
+@RequestMapping("/itens")
+public class ItemController {
+
+    @Autowired
+    private ItemService itemService;
+
+    @GetMapping
+    public ResponseEntity<List<Item>> listarItens() {
+        List<Item> itens = itemService.listarItens();
+        return new ResponseEntity<>(itens, HttpStatus.OK);
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<Item> encontrarItemPorId(@PathVariable int id) {
+        return itemService.encontrarItemPorId(id)
+                .map(item -> new ResponseEntity<>(item, HttpStatus.OK))
+                .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
+    }
+
+    @PostMapping
+    public ResponseEntity<Item> salvarItem(@RequestBody Item item) {
+        Item novoItem = itemService.salvarItem(item);
+        return new ResponseEntity<>(novoItem, HttpStatus.CREATED);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deletarItem(@PathVariable int id) {
+        itemService.deletarItem(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
+}
