+++ src/main/java/br/com/pedrolg/projeto/controller/PedidoController.java
@@ -1,0 +1,50 @@
+package br.com.pedrolg.projeto.controller;
+
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import br.com.pedrolg.projeto.entity.Pedido;
+import br.com.pedrolg.projeto.service.PedidoService;
+
+@RestController
+@RequestMapping("/pedidos")
+public class PedidoController {
+
+    @Autowired
+    private PedidoService pedidoService;
+
+    @GetMapping
+    public ResponseEntity<List<Pedido>> listarPedidos() {
+        List<Pedido> pedidos = pedidoService.listarPedidos();
+        return new ResponseEntity<>(pedidos, HttpStatus.OK);
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<Pedido> encontrarPedidoPorId(@PathVariable Long id) {
+        return pedidoService.encontrarPedidoPorId(id)
+                .map(pedido -> new ResponseEntity<>(pedido, HttpStatus.OK))
+                .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
+    }
+
+    @PostMapping
+    public ResponseEntity<Pedido> salvarPedido(@RequestBody Pedido pedido) {
+        Pedido novoPedido = pedidoService.salvarPedido(pedido);
+        return new ResponseEntity<>(novoPedido, HttpStatus.CREATED);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deletarPedido(@PathVariable Long id) {
+        pedidoService.deletarPedido(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
+}
