+++ src/main/java/br/com/pedrolg/projeto/repository/ItemRepository.java
@@ -1,0 +1,10 @@
+package br.com.pedrolg.projeto.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import br.com.pedrolg.projeto.entity.Item;
+
+@Repository
+public interface ItemRepository extends JpaRepository<Item, Integer> {
+}
