+++ src/main/java/br/com/pedrolg/projeto/service/ClienteService.java
@@ -1,0 +1,33 @@
+package br.com.pedrolg.projeto.service;
+
+import java.util.List;
+import java.util.Optional;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import br.com.pedrolg.projeto.entity.Cliente;
+import br.com.pedrolg.projeto.repository.ClienteRepository;
+
+@Service
+public class ClienteService {
+
+    @Autowired
+    private ClienteRepository clienteRepository;
+
+    public List<Cliente> listarClientes() {
+        return clienteRepository.findAll();
+    }
+
+    public Optional<Cliente> encontrarClientePorId(int id) {
+        return clienteRepository.findById(id);
+    }
+
+    public Cliente salvarCliente(Cliente cliente) {
+        return clienteRepository.save(cliente);
+    }
+
+    public void deletarCliente(int id) {
+        clienteRepository.deleteById(id);
+    }
+}
