+++ src/main/java/br/com/pedrolg/projeto/service/PedidoService.java
@@ -1,0 +1,37 @@
+package br.com.pedrolg.projeto.service;
+
+import java.util.List;
+import java.util.Optional;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import br.com.pedrolg.projeto.entity.Pedido;
+import br.com.pedrolg.projeto.repository.ItemRepository;
+import br.com.pedrolg.projeto.repository.PedidoRepository;
+
+@Service
+public class PedidoService {
+
+    @Autowired
+    private PedidoRepository pedidoRepository;
+
+    @Autowired
+    private ItemRepository itemRepository;
+
+    public List<Pedido> listarPedidos() {
+        return pedidoRepository.findAll();
+    }
+
+    public Optional<Pedido> encontrarPedidoPorId(Long id) {
+        return pedidoRepository.findById(id);
+    }
+
+    public Pedido salvarPedido(Pedido pedido) {
+        return pedidoRepository.save(pedido);
+    }
+
+    public void deletarPedido(Long id) {
+        pedidoRepository.deleteById(id);
+    }
+}
