--- src/main/java/myapp/grupocarecaspring/GrupocarecaspringApplication.java
@@ -1,13 +1,0 @@
-package myapp.grupocarecaspring;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class GrupocarecaspringApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(GrupocarecaspringApplication.class, args);
-	}
-
-}
