--- src/main/java/myapp/grupocarecaspring/controllers/ClienteController.java
@@ -1,44 +1,0 @@
-package myapp.grupocarecaspring.controllers;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import myapp.grupocarecaspring.entities.Cliente;
-import myapp.grupocarecaspring.services.ClienteService;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/clientes")
-public class ClienteController {
-
-    @Autowired
-    private ClienteService clienteService;
-
-    @GetMapping
-    public ResponseEntity<List<Cliente>> listarClientes() {
-        List<Cliente> clientes = clienteService.listarClientes();
-        return new ResponseEntity<>(clientes, HttpStatus.OK);
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<Cliente> encontrarClientePorId(@PathVariable int id) {
-        return clienteService.encontrarClientePorId(id)
-                .map(cliente -> new ResponseEntity<>(cliente, HttpStatus.OK))
-                .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
-    }
-
-    @PostMapping
-    public ResponseEntity<Cliente> salvarCliente(@RequestBody Cliente cliente) {
-        Cliente novoCliente = clienteService.salvarCliente(cliente);
-        return new ResponseEntity<>(novoCliente, HttpStatus.CREATED);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deletarCliente(@PathVariable int id) {
-        clienteService.deletarCliente(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-    }
-}
