--- src/main/java/myapp/grupocarecaspring/controllers/PedidoController.java
@@ -1,44 +1,0 @@
-package myapp.grupocarecaspring.controllers;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import myapp.grupocarecaspring.entities.Pedido;
-import myapp.grupocarecaspring.services.PedidoService;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/pedidos")
-public class PedidoController {
-
-    @Autowired
-    private PedidoService pedidoService;
-
-    @GetMapping
-    public ResponseEntity<List<Pedido>> listarPedidos() {
-        List<Pedido> pedidos = pedidoService.listarPedidos();
-        return new ResponseEntity<>(pedidos, HttpStatus.OK);
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<Pedido> encontrarPedidoPorId(@PathVariable Long id) {
-        return pedidoService.encontrarPedidoPorId(id)
-                .map(pedido -> new ResponseEntity<>(pedido, HttpStatus.OK))
-                .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
-    }
-
-    @PostMapping
-    public ResponseEntity<Pedido> salvarPedido(@RequestBody Pedido pedido) {
-        Pedido novoPedido = pedidoService.salvarPedido(pedido);
-        return new ResponseEntity<>(novoPedido, HttpStatus.CREATED);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deletarPedido(@PathVariable Long id) {
-        pedidoService.deletarPedido(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-    }
-}
