--- src/main/java/myapp/grupocarecaspring/repositories/ClienteRepository.java
@@ -1,10 +1,0 @@
-package myapp.grupocarecaspring.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import myapp.grupocarecaspring.entities.Cliente;
-
-@Repository
-public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
-}
