--- src/main/java/myapp/grupocarecaspring/repositories/PedidoRepository.java
@@ -1,10 +1,0 @@
-package myapp.grupocarecaspring.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import myapp.grupocarecaspring.entities.Pedido;
-
-@Repository
-public interface PedidoRepository extends JpaRepository<Pedido, Long> {
-}
